Add and enable Azure support by default; helps build order
In pursuing a manylinux2010 based build the openssl library lookups were
failing when building cURL; likely because of misuse of the
`--with-ssl` configure flag.  As currently structured the configure
generated cURL Makefile does not correctly add multilib paths to the
linker search paths due to pkgconfig metadata not being used.
Specifying `with-ssl` with a path creates this scenario.  If
openssl can find multilib [1] directories during the configure
phase of the build it will use them.  Currently, via cmake in
this repository, TileDB will be built in such a way that openssl
will be built after the `lib64` multilib directory has been created.

TileDB-Py, which also uses a manylinux2010 build, has these same
structural problems but includes Azure support by default.  This
changes the external dependency build order, openssl comes up first,
the `lib64` directoriy is missing `LIBDIR` is set to `lib` and the
rest of the build completes successfully.

Not wanting to hack on the TileDB build itself the easiest solution
was to add Azure support to TileDB-Java.  It also has the nice side
effect of including the support.

 1. https://github.com/openssl/openssl/blob/OpenSSL_1_1_0h/Configurations/unix-Makefile.tmpl#L147-L155