Update SONATYPE publishing
This updates the SONATYPE url for the new staging API that serves as a shim.
We also update to new ENVs for username/password for maven upload. I opted for
new secrets in the repo, so I wouldn't override the old secrets while setting
this up. The old secrets will be safe to delete after merging.